# perf: remove ts resolved module cache file