## Summary

I noticed that running ts-jest was very slow in one of my projects so I dig a little into why and noticed the performance bottleneck comes from the `getCacheKey` function and in particular the `writeFileSync` that is being called to store the dependency graph.

Writing this file is what's taking most of the time in this process as the content we write is very heavy : the whole file content and its resolved module names.

I ran a few tests on a project with different configurations and here are the result:
| Case | 1st run (clean jest cache) | 2nd run (with jest cache)
|---|----|----|
|Current ts-jest (Writing dependency graph file) | 218s | 16s|
|No cache (`--no-cache`) | 44s | 44s|
|This PR version (No dependency graph file) | 47s | 14s|

As we can see from the results, the fastest way to run ts-jest is by disabling jest cache (`--no-cache`) but we can get pretty close to that timing when we skip writing the dep graph file. 

What's even more interesting is that on subsequent runs, even reading the dep graph file is less performant than recomputing the graph (2s less in this case)

So I'm opening this PR to completely remove the dependency graph file as it seems to only slow down the process without any gain. It should largely improve ts-jest performance when not using isolated modules. 

## Test plan

<!-- Demonstrate the code is solid. Example: The exact commands you ran and their output, screenshots / videos if the pull request changes UI. -->

## Does this PR introduce a breaking change?

- [ ] Yes
- [x] No


## Other information