@ahnpnl 2024-05-19
In the past, we introduced that checking dep graph and includes related files into cache key because we had the issue in watch mode.

You can check https://github.com/kulshekhar/ts-jest/pull/2436 for more information.

Reverting that PR is fine, it will solve problems for some users but introduce past problems for other users. 

I haven’t found a way how to do type checking probably either. Apparently doing type checking inside Jest transformer is very poor performant. If we can somehow do it from top level before Jest runs, it would be ideal.