@jsaguet-betclic 2024-05-19
The dep grah itself and how it's computed is not removed or changed with this PR.
It will just be recomputed every time the cache key is computed.
The only change is that it will not be stored between runs because that's the part that slows ts-jest a lot.