@decline 2024-05-21
Just wanted to leave a big THANK YOU for this fix! We have a very large codebase where we couldn't even run all tests at once anymore, without setting `isolatedModules: true` because of the long run times. Now everything is super fast and we can have type-safety again by removing the `isolatedModules` setting. Really happy 🎉